Allow more recent versions of dependencies
In general, Python libraries should not depend on fixed versions of their dependencies. Update requirements.txt to set a minimum version of the packages we depend on but no upper limit. This will allow pysgconnect to be installed as a dependency of applications that use more recent versions of these packages.